Extract base class for xtext file creation wizards.
